tidy up code, switch to manual promote to prod and override nexus admin password

+++ pkg/config/helm_values.go
@@ -18,11 +18,6 @@ type ExposeController struct {
 	Annotations map[string]string      `yaml:"Annotations,omitempty"`
 }
 
-type Annotation struct {
-	Hook   string `yaml:"helm.sh/hook,omitempty"`
-	Delete string `yaml:"hhelm.sh/hook-delete-policy,omitempty"`
-}
-
 type HelmValuesConfig struct {
 	ExposeController *ExposeController `yaml:"expose,omitempty"`
 }
@@ -36,13 +31,17 @@ func (c *HelmValuesConfig) AddExposeControllerValues(cmd *cobra.Command, ignoreD
 	if !ignoreDomain {
 		cmd.Flags().StringVarP(&c.ExposeController.Config.Domain, "domain", "", "", "Domain to expose ingress endpoints.  Example: jenkinsx.io")
 	}
+	keepJob := false
 	cmd.Flags().BoolVarP(&c.ExposeController.Config.HTTP, "http", "", true, "Toggle creating http or https ingress rules")
 	cmd.Flags().StringVarP(&c.ExposeController.Config.Exposer, "exposer", "", "Ingress", "Used to describe which strategy exposecontroller should use to access applications")
 	cmd.Flags().BoolVarP(&c.ExposeController.Config.TLSAcme, "tls-acme", "", false, "Used to enable automatic TLS for ingress")
+	cmd.Flags().BoolVarP(&keepJob, "keep-exposecontroller-job", "", false, "Prevents Helm deleting the exposecontroller Job and Pod after running.  Useful for debugging exposecontroller logs but you will need to manually delete the job if you update an environment")
 
 	annotations := make(map[string]string)
 	annotations["helm.sh/hook"] = "post-install,post-upgrade"
-	annotations["helm.sh/hook-delete-policy"] = "hook-succeeded"
+	if !keepJob {
+		annotations["helm.sh/hook-delete-policy"] = "hook-succeeded"
+	}
 	c.ExposeController.Annotations = annotations
 }
